refactor: move auth logic and view to users app
