+++ notes/views.py
@@ -1,40 +1,10 @@
-from django.contrib.auth import authenticate, login
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse
 from django.http import HttpResponseNotFound
 from django_ratelimit.decorators import ratelimit
 from .models import Note
-from .forms import NoteForm, RegistrationForm, LoginForm
-
-
-@ratelimit(key="ip", rate="50/h", block=True)
-def register_view(request):
-    if request.method == "POST":
-        form = RegistrationForm(request.POST)
-        if form.is_valid():
-            user = form.save()
-            login(request, user)
-            return redirect(reverse("notes:create_note"))
-    else:
-        form = RegistrationForm()
-    return render(request, "notes/register.html", {"form": form})
-
-
-@ratelimit(key="ip", rate="50/h", block=True)
-def login_view(request):
-    if request.method == "POST":
-        form = LoginForm(request.POST)
-        if form.is_valid():
-            username = form.cleaned_data["username"]
-            password = form.cleaned_data["password"]
-            user = authenticate(request, username=username, password=password)
-            if user is not None:
-                login(request, user)
-                return redirect(reverse("notes:list_notes"))
-    else:
-        form = LoginForm()
-    return render(request, "notes/login.html", {"form": form})
+from .forms import NoteForm
 
 
 @login_required
@@ -63,7 +33,7 @@ def note_detail_view(request, pk):
     if note.is_expired():
         note.delete()
         return HttpResponseNotFound(
-            "This note has expired or been viewed the maximum number of times."
+            "this note has expired or been viewed the maximum number of times."
         )
     note.current_views += 1
     note.save()
