+++ secretnote/urls.py
@@ -20,5 +20,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path("notes/", include("notes.urls")),
+    path("", include("users.urls")),
     path("admin/", admin.site.urls),
 ]
